feat(compleo_wallbox): Add support for RFID tag reading
The commit adds support for reading the RFID tag from the Compleo Wallbox device. This feature allows users to identify the vehicle connected to the charger using the RFID tag. The change includes adding a new register to the constant definitions and updating the code to retrieve and expose the RFID tag information.